refactor: 💡 UsersDb class moved to the ./node/db directory
