+++ src/node/api/index.ts
@@ -1,4 +1,4 @@
-export * from './db.js';
+export * from '../db/users.js';
 export * from './server.js';
 export * from './router/admin.js';
 export * from './router/user.js';
